aider: Update `appendSystemMessage` function in `static\script.js` to format the text and display special characters properly as HTML, including double newlines.
# Aider chat conversation:

USER: format the responses from the assistant. If there are any \n \r \t eccetera, make sure that it's reflected in the frontend. There shouldn't be any of this characters displayed, just html code. If there is a double \n\n insert a double <br>

ASSISTANT: To format the responses from the assistant in the frontend, we will modify the `appendSystemMessage` function in the `static\script.js` file. This function will be updated to replace newline (`\n`), carriage return (`\r`), and tab (`\t`) characters with their corresponding HTML tags or entities. Additionally, if there is a double `\n\n`, it will be replaced with a double `<br><br>`.

Here is the *SEARCH/REPLACE* block for `static\script.js`:

static\script.js
```javascript
<<<<<<< SEARCH
        // Format the text to display special characters properly as HTML
        const formattedText = text
            .replace(/\n/g, '<br>')
            .replace(/\r/g, '') // Carriage returns are not needed in HTML, remove them
            .replace(/\t/g, '&nbsp;&nbsp;&nbsp;&nbsp;'); // Replace each tab with 4 non-breaking spaces
=======
        // Format the text to display special characters properly as HTML, including double newlines
        const formattedText = text
            .replace(/\n\n/g, '<br><br>') // Replace double newlines with double line breaks
            .replace(/\n/g, '<br>') // Then replace remaining single newlines with line breaks
            .replace(/\r/g, '') // Carriage returns are not needed in HTML, remove them
            .replace(/\t/g, '&nbsp;&nbsp;&nbsp;&nbsp;'); // Replace each tab with 4 non-breaking spaces
>>>>>>> REPLACE
```

This change ensures that double newlines are converted to double `<br>` tags before converting single newlines to single `<br>` tags, and other special characters are handled as before.